bigint fix1 1. sign_ now int8_t instead of int 2. cycles kinda    for (size_t i = n; i != 0;) {        --i;        ...    }    replaces with    for (size_t i = n; i-->0;) {...} 3. changed sign formula from (a > b) - (a < b) to 1 - ((a < b) << 1U)

+++ bigint/big_integer.h
@@ -57,7 +57,7 @@ private:
     void resize_(size_t n);
 
 private:
-    int sign_;
+    int8_t sign_;
     storage_t words_;
 };
